+++ HISTORY.rst
@@ -1,6 +1,9 @@
 =======
 History
 =======
+2025.2.9 -- Added bond lentghs, angles, etc. to results
+    * This allows them to be saved to variables, JSON, etc. for external analysis or use
+      in the flowchart.
 
 2022.12.24 (2022-12-24)
 -----------------------